refactor: 모달 외부 클릭 감지 로직을 useClickOutside 훅으로 분리 후 UserMenuModal에 적용
