+++ src/hooks/useClickOutside.ts
@@ -1,0 +1,22 @@
+import { useEffect, useRef } from 'react';
+
+export const useClickOutside = (onClose: () => void) => {
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        onClose();
+      }
+    };
+
+    if (onClose !== undefined) {
+      document.addEventListener('mousedown', handleClickOutside);
+      return () => {
+        document.removeEventListener('mousedown', handleClickOutside);
+      };
+    }
+  }, [onClose]);
+
+  return ref as React.RefObject<HTMLDivElement>;
+};
